move SKRWebhookConfig to deploy package
